feat: filters on the tables in dashboard

+++ assurapp-web/src/app/components/dashboard/insurances-dashboard/insurances-dashboard.component.ts
@@ -7,18 +7,25 @@ import {AuthenticationService} from "../../../services/authentication.service";
 import {UserService} from "../../../services/user.service";
 import {PopupService} from "../../../services/popup.service";
 import {PopupType} from "../../../enums/popup-type";
+import {SubscriptionService} from "../../../services/subscription.service";
+import {PaginatorModule} from "primeng/paginator";
+import {ReactiveFormsModule} from "@angular/forms";
+import {InsuranceType} from "../../../enums/insurance-type";
 
 @Component({
   selector: 'app-insurances-dashboard',
   standalone: true,
   imports: [
     RouterLink,
-    CurrencyPipe
+    CurrencyPipe,
+    PaginatorModule,
+    ReactiveFormsModule
   ],
   templateUrl: './insurances-dashboard.component.html',
   styleUrl: './insurances-dashboard.component.css'
 })
 export class InsurancesDashboardComponent implements OnInit {
+  types:InsuranceType[] = Object.values(InsuranceType);
   insurances: Insurance[] = [];
   data: Insurance[] = [];
   current = 1;
@@ -26,11 +33,20 @@ export class InsurancesDashboardComponent implements OnInit {
 
   constructor(private insuranceService: InsuranceService,
               private authenticationService: AuthenticationService,
+              private subscriptionService: SubscriptionService,
               private userService: UserService,
               private popupService: PopupService) {}
 
   ngOnInit(): void {
     this.fetch();
+  }
+
+  filter(event: any) {
+    if (event.target.value === "ALL") {
+      this.getData();
+      return
+    }
+    this.data = this.insurances.filter(i => i.type === event.target.value);
   }
 
   fetch() {
@@ -49,11 +65,20 @@ export class InsurancesDashboardComponent implements OnInit {
   }
 
   delete(id: number) {
-    if (!confirm("Are you sure to delete this insurance")) return;
-    this.insuranceService.deleteInsurance(id).subscribe({
-      next: data => this.fetch(),
-      error: () =>  this.popupService.show("Can't delete to API", PopupType.ERROR)
-    });
+    this.subscriptionService.getSubscriptionByInsurer(this.authenticationService.getUserId()).subscribe({
+      next: data => {
+        if (data.some(s => s.insurance.id == id && new Date(s.endDate) > new Date()))
+          this.popupService.show("There is subscriptions linked to this insurance.", PopupType.WARNING)
+        else {
+          if (!confirm("Are you sure to delete this insurance")) return;
+          this.insuranceService.deleteInsurance(id).subscribe({
+            next: data => this.fetch(),
+            error: () =>  this.popupService.show("Can't delete to API", PopupType.ERROR)
+          });
+        }
+      },
+      error: () => this.popupService.show("Can't delete to API", PopupType.ERROR)
+    })
   }
 
   getData() {
